fix: 메서드 반환타입 일치 (#481)
* fix: 메서드 반환타입 일치

* fix: 반환 타입을 Long 으로 변경

* fix: 반환 순서 및 타입 일치

* chore: application count 타입도 long 으로 통일

+++ src/main/java/com/becareful/becarefulserver/domain/caregiver/dto/response/CaregiverHomeResponse.java
@@ -9,16 +9,16 @@ import lombok.Builder;
 public record CaregiverHomeResponse(
         String name,
         boolean hasNewChat,
-        Integer applicationCount,
-        Integer recruitmentCount,
+        Long applicationCount,
+        Long recruitmentCount,
         List<WorkScheduleResponse> workScheduleList,
         boolean isWorking,
         boolean isApplying) {
     public static CaregiverHomeResponse of(
             Caregiver caregiver,
             boolean isNewChat,
-            Integer applicationCount,
-            Integer recruitmentCount,
+            Long applicationCount,
+            Long recruitmentCount,
             boolean isWorking,
             boolean isApplying,
             List<WorkScheduleResponse> workScheduleList) {
