+++ src/main/java/com/becareful/becarefulserver/domain/matching/repository/ApplicationRepository.java
@@ -42,5 +42,5 @@ public interface ApplicationRepository extends JpaRepository<Application, Long> 
 
     boolean existsByRecruitment(Recruitment recruitment);
 
-    Integer countByWorkApplication(WorkApplication workApplication);
+    Long countByWorkApplication(WorkApplication workApplication);
 }
